requirements.txt


Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
#matplotlib==3.0.3
numpy==1.16.2

pandas==0.24.2
pandocfilters==1.4.2

pickleshare==0.7.5
Pillow==6.0.0


scikit-learn==0.20.3

scipy==1.2.1


tensorflow-gpu==1.12.0
opencv-python==3.4.2.17

